Create storage support for EmptyUser
-Split User into ExistingUser and EmptyUser
