+++ src/main/java/seedu/address/model/person/user/User.java
@@ -16,102 +16,6 @@ import java.util.Set;
 import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
 
 /**
- * Represents the User of the address book.
- * Guarantees: details are present and not null, field values are validated, immutable.
+ * Represents the User of the address book. User can never be null.
  */
-public class User {
-
-    // Identity fields
-    private final Name name;
-    private final Phone phone;
-    private final Email email;
-
-    // Data fields
-    private final Address address;
-    private final Set<CurrentModule> currModules = new HashSet<>();
-    private final Set<PreviousModule> prevModules = new HashSet<>();
-    private final Set<PlannedModule> planModules = new HashSet<>();
-
-    /**
-     * Every field must be present and not null.
-     */
-    public User(Name name, Phone phone, Email email, Address address, Set<CurrentModule> currModules,
-                Set<PreviousModule> prevModules, Set<PlannedModule> planModules) {
-        requireAllNonNull(name, phone, email, address);
-        this.name = name;
-        this.phone = phone;
-        this.email = email;
-        this.address = address;
-        this.currModules.addAll(currModules);
-        this.prevModules.addAll(prevModules);
-        this.planModules.addAll(planModules);
-    }
-
-    public User() {
-        this.name = new Name("emptyUser");
-        this.phone = new Phone("000");
-        this.email = new Email("empty@user.com");
-        this.address = new Address("emptyUser");
-    }
-
-    public Name getName() {
-        return name;
-    }
-
-    public Phone getPhone() {
-        return phone;
-    }
-
-    public Email getEmail() {
-        return email;
-    }
-
-    public Address getAddress() {
-        return address;
-    }
-
-    /**
-     * Returns an immutable current module set, which throws {@code UnsupportedOperationException}
-     * if modification is attempted.
-     */
-    public Set<CurrentModule> getCurrModules() {
-        return Collections.unmodifiableSet(currModules);
-    }
-
-    /**
-     * Returns an immutable previous module set, which throws {@code UnsupportedOperationException}
-     * if modification is attempted.
-     */
-    public Set<PreviousModule> getPrevModules() {
-        return Collections.unmodifiableSet(prevModules);
-    }
-
-    /**
-     * Returns an immutable planned module set, which throws {@code UnsupportedOperationException}
-     * if modification is attempted.
-     */
-    public Set<PlannedModule> getPlanModules() {
-        return Collections.unmodifiableSet(planModules);
-    }
-
-    @Override
-    public int hashCode() {
-        // use this method for custom fields hashing instead of implementing your own
-        return Objects.hash(name, phone, email, address);
-    }
-
-    @Override
-    public String toString() {
-        final StringBuilder builder = new StringBuilder();
-        builder.append(getName())
-                .append("; Phone: ")
-                .append(getPhone())
-                .append("; Email: ")
-                .append(getEmail())
-                .append("; Address: ")
-                .append(getAddress());
-
-        return builder.toString();
-    }
-
-}
+public abstract class User {}
